requirements.txt
streamlit==1.32.0
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
openai>=1.68.2,<2.0.0
python-dotenv==1.0.1
streamlit-option-menu==0.3.12
plotly==5.19.0
pandas==2.2.0
altair==5.2.0
transformers==4.38.2
torch==2.2.1
PyMuPDF==1.23.26
python-docx==1.1.0
requests==2.31.0
lottie==0.7.2
pydantic==2.11.4
